feat: 웹에서 코드 입력받기
