--- src/js/component.js
@@ -1,3 +1,0 @@
-export function TaskComponent(task) {
-  return `<span class="task">${task}</span>`;
-}
