+++ src/js/main.js
@@ -1,21 +1,8 @@
-import {
-  registerToWebAPI,
-  classifyIntoMacroAndMicro,
-  moveToQueue,
-  moveToCallstack,
-} from './controller.js';
-import { parseCode } from './parsing.js';
-import { code } from './test-code.js';
+import { getCode } from './input.js';
+// import { code } from './test-code.js';
 
 async function main() {
-  const tasks = parseCode(code);
-  console.log(tasks);
-  await registerToWebAPI(tasks);
-  const { microTask, macroTask } = classifyIntoMacroAndMicro(tasks);
-  await moveToQueue(microTask);
-  await moveToQueue(macroTask);
-  await moveToCallstack(microTask);
-  moveToCallstack(macroTask);
+  getCode();
 }
 
 main();
